Move flushPromises into testUtils file to avoid repeated mocks during import #1379
