+++ src/state/middleware/scigateway.middleware.test.tsx
@@ -25,7 +25,7 @@ import { toastr } from 'react-redux-toastr';
 import { AddHelpTourStepsType } from '../scigateway.types';
 import { StateType } from '../state.types';
 import TestAuthProvider from '../../authentication/testAuthProvider';
-import { flushPromises } from '../../setupTests';
+import { flushPromises } from '../../testUtils';
 import { authState, initialState } from '../reducers/scigateway.reducer';
 import { buildTheme } from '../../theming';
 import { thunk } from 'redux-thunk';
