+++ src/testUtils.tsx
@@ -1,0 +1,2 @@
+export const flushPromises = async (): Promise<unknown> =>
+  (await vi.importActual('timers')).setImmediate;
